pyproject.toml
[tool.poetry]
name = "data-hopper"
version = "0.1.0"
description = ""
authors = ["tomasoak <tomas.jpeg@gmail.com>"]
readme = "README.md"
packages = [{include = "datahopper"}]

[tool.poetry.dependencies]
python = "^3.8"
attrs = "^22.1.0"
iniconfig = "^1.1.1"
numpy = "^1.23.3"
packaging = "^21.3"
pandas = "^1.5.0"
pluggy = "^1.0.0"
py = "^1.11.0"
pyparsing = "^3.0.9"
pytest = "^7.1.3"
python-dateutil = "^2.8.2"
pytz = "^2022.2.1"
six = "^1.16.0"
tomli = "^2.0.1"
mkdocs-material = "^8.5.3"
Unidecode = "^1.3.6"
plotly-express = "^0.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
